requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
TikTokApi>=7.0.0
yt-dlp>=2023.7.6
openai>=1.0.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
nest_asyncio>=1.5.6
# pyppeteer>=1.0.0  # 改為使用原始JavaScript爬蟲
astrapy>=0.5.0
langchain-openai>=0.0.3
requests>=2.28.0
cloudinary>=1.36.0
mangum
Pillow>=9.0.0